cc26xx_cc13xx: fix VIMS/FLASH docs and offsets
For the `VIMS` and `FLASH` a value of 0x4 was being added,
I couldn't find a reason for that.

- Base address for VIMS is 0x40034000, and the STAT and CTL registers
are defined as follows on the TI SDK:

    #define VIMS_O_STAT 0x00000000
    #define VIMS_O_CTL  0x00000004

So `CTL` was incorrectly being mapped onto `STAT` in the `vims_regs_t`
structure.

- For FLASH the base address is 0x40030000, the first register is
defined as, so there's a gap of 0x1C which is represented by the
`__reserved0` field on `flash_regs_t` structure:

    #define FLASH_O_STAT 0x0000001C

However the 0x4 was being added to shifting incorrectly the registers.

Signed-off-by: Jean Pierre Dudey <jeandudey@hotmail.com>