tests/gnrc_netif_delayed_start: add test for delayed start of interfaces
This tests application defined startup & shutdown of network interfaces.

Battery powered devices might only want to enable the radio if they have
data to send.
For the most power saving and least startup time, that means they will not
even configure the radio when not needed and just leave it in Deep Sleep.

For this to work the radio must remain in Deep Sleep across reboots and also
allow to be shut down properly.

This adds a test to exercise that use-case.