tests/driver_dac_dds: add test for dac_dds_play()
This will play a raw audio (8kHz, 8 Bit) snipped from the 1961 synthesized speech demo.
(Can be disabled since the file is rather large with 17k)

It will also play sine, square and sawtooth waves with frequencies that can be set on
the command line.

I tested this with the on-board speaker (connected to P0.26) of the mcb2388 as well
as with hooking up a headphone to the ANALOG (PA02) header of the same54-xpro.