cpu/esp32: switch from O2 to Os
In #12955 optimization was switche to O2 because with the '-Os'
option, the ESP32 hangs sporadically in 'tests/bench*' if
interrupts where disabled too early by benchmark tests.

Since hasn't been reproduced since and in #13196 O2 was causing
un-explained hardfaults, since the formentioned issue could not
be reproduced we switch back to Os.