examples/micropython: fix checked PID
With https://github.com/RIOT-OS/RIOT/pull/14224 the idle thread became
optional, so the main thread can have either 1 or 2 as PID, depending
if the idle thread is included or not. As this might also change in the
future, it is probably the best to just expect any number.

(cherry picked from commit 5599b067abd6a1d61ab64aa286f484956aa52804)