drivers/cc2420: uncomment bogus cc2420_set_state(dev, CC2420_GOTO_RX)
It doesn't compile, compiler complains that the enum type of
CC2420_GOTO_RX cannot be converted to netopt_t as required by the
function.
The numerical value of CC2420_GOTO_RX would be NETOPT_ADDRESS (2), which
can't be right.
cc2420_set_state(dev, NETOPT_STATE_RX) in turn is not implemented
(returns -ENOTSUP), so this was a noop for a while.
TODO: revisit and confirm.