boards: add stm32mp157c-dk2 initial support
This board is based on a stm32mp157cac which has a dual architecture:
* Dual core Cortex-A7
* Cortex-M4
Only Cortex-M4 is supported by RIOT-OS.
Cortex-M4 can be used in Engineering mode if stm32mp1_eng_mode
pseudomodule is used.
By default the RIOT firmware can be loaded by Linux on the Cortex-M4
using remoteproc Linux framework.
This the initial commit with a limited set of supported peripheral:
* gpio
* timer
* uart

Signed-off-by: Gilles DOFFE <gilles.doffe@savoirfairelinux.com>