socket_zep: fix reboot
The `_zep_params_setup()` function will modify the `argv` string passed to it.
This is a problem because that string is re-used on reboot.

The modified string is then later processed in `socket_zep_setup()`, so we have to keep
the memory around.

The `strdup()` fulfills all this with the only downside that we now have a minor memory
leak on every native reboot with ZEP.

IMHO as this is a pretty rare event and on `native` there is plenty of memory, a few
bytes lost are not a big deal.