drivers/saul/auto_init: Support bool and percent for dimmers
Booleans make sense for dimmers when accessed from a Sith application
that deals in absolutes.

Percent values are also a widespread interpretation of brightness
levels, and are thus supported as well.

The bit arithmetic makes sure that the arithmetic operation value / 100
* saul_pwm_resolution is done efficiently (by expression as a
multiplication followed by shifting) and accurately (by maximizing the
number of usable bits) while still being flexible both with respect to
integer sizes and to changes of saul_pwm_resolution.