fixup! drivers/saul/auto_init: Add PWM for LEDs
Change int to uint to avoid needless casting, and to align with the
printed type used.

On the downside, this now necessitates an extra assignement to a
constant not declared as such, for otherwise Werror=type-limits would
reject the never-entered loop.

(A previous workaround using #ifdef was rejected due to its fall-out:
then arguments become unused, then statics ...)