boards/bluepill-stm32f030c8: add STM32F030C8 based bluepill board
This board is sold by several vendors on AliExpress, Amazon and ebay.
It does not have name, but since it comes in the same form-factor and
color as the Bluepill line of boards, I think it makes sense to use that
name.

Flashing is a bit cumbersome as I always had to press the reset button
in time when OpenOCD tries to connect to the CPU.
There is no dedicated reset pin exposed on the board.