core/mutex: Add mutex_cancel()
This commit adds the (pseudo)-module `mutex_cancel`, which unlocks the likewise
named function. With this function, a thread blocked with `mutex_lock()` can
be unblocked without it obtaining the mutex. A caller of `mutex_lock()` is
notified via an `-ECANCELED` return code.

This is intended to simplify the implementation of `xtimer_mutex_lock_timeout()`
and to implement `ztimer_mutex_lock_timeout()`.