boards/qn9080dk: Add mma8x5x driver for MMA8652FC
The qn9080dk includes a MMA8652FC chip connected to the I2C0 bus, this
patch adds the driver for the in the default.

Note that the MMA8652FC device is powered from the LPC_VCC line, which
is different from the QN_VCC line that powers the QN9080 MCU. A voltage
level translator is included in the QN9080DK which is only enabled when
JP11 (labeled "G-Sensor I2C") is closed and the LPC debugger hardware
is powered, for example from the USB port.