sys/net/gnrc/sock: cleanup & fix aux handling
The logic used to check whether the RX timestamp was send in the GNRC
implementation of `sock_ip_recv_buf_aux()` is incorrect: It still uses in-band
signalling via a timestamp of zero, but a dedicated flag was added to allow for
timestamps of zero.

Additionally the GNRC implementations of both `sock_ip_recv_buf_aux()` and
`sock_udp_recv_buf_aux()` have been simplified: There is little value to check
if a bit is set an than clearing it - one can just as well simply clear it
unconditionally.