rtt_rtc: make use of RTT overflow callback when available
Some RTT implementations provide a separate overflow callback.
`rtt_rtc` was originally designed to work with the most basic RTT:
It only needs a monotonic counter and an alarm callback that is shared
between overflow and RTC alarm.

If we use the overflow callback, this frees up the alarm callback for
direct use with the RTT API.

This means that if an RTT provides the alarm callback and the user treads it
as a monotonic counter (no calls to `rtt_set_counter()`), both `rtc_rtt` and
the `periph_rtt` API can be used together.

The user also must make sure to either use RTC or RTT alarms as they will overwrite
each other.

But this allows to track & adjust the time without having to re-calculate RTT alarms
if the time is changed due to synchronisation.