driver/bmx280: increase accuracy of SAUL pressure readings
The SAUL implementation of the BMx280 currently returns the pressure in hPa.
The sensor, however, actually reads the pressure in Pa, leading to a
significant loss of precision.

Because the operation range of the BME280 and BMP280 is 300 hPa to 1100 hPa,
we can safely increase the precision by a factor of 10 without risk of overflow.
The only downside is that readings are now presented as `e1 Pa` (eg 10206e1 Pa).