cpu/avr8_common: Enable avr-8 nested interrupts
AVR-8 supports nested interrupts but there is no infrastructure to
use that.  This refactor infraestrocuture and enable the enhancement.

The global avr8_state was removed and responsability was split into
two general purpose IO Registers:

 - GPIOR0: store how deep is the nested interrupt level.
 - GPIOR1: indicate when uart tx is in pending state.

The GPIOR0 have the function to flag when system is processing a ISR.
At end of ISR routine, scheduler may require a switch context.  This
can be performed only when GPIOR0 is zero.  Otherwise process stack
may be corrupted.

The GPIOR1 continue with same meaning as before.  The big difference
now is, it supports up to 8 uarts and there are dedicated methods to
set/unset tx pending.

Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>