cpu/avr8_common: Improve isr state for nested interrupts
ATxmega supports nested interrupts but there is no infrastructure
to use that.  Split avr8_state into two new variables: avr8_isr_state
and avr8_uart_state.

The avr8_isr_state is a count responsible to store how deep is the
nested interrupt level.  The scheduler need check if deep is zero to
perform a switch context.

The avr8_uart_state continue with same meaning before.  The big
difference now is, it supports up to 8 uarts and there is a dedicated
method to set/unset tx pending.

Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>